requirements.txt
absl-py==2.3.1
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.7.14
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
ipykernel==6.30.0
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter-client==8.6.3
jupyter-core==5.8.1
keras==3.11.0
kiwisolver==1.4.8
lazy-loader==0.4
libclang==18.1.1
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.5.3
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.1.3
nvidia-nccl-cu12==2.27.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
rich==14.1.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
simple-image-download==0.2
six==1.17.0
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.2
